refactor env loading and simplify tests
