--- tests/assert.go
@@ -1,17 +1,0 @@
-package tests
-
-import (
-	"io"
-	"net/http"
-	"testing"
-)
-
-func readAllAndClose(t *testing.T, resp *http.Response) []byte {
-	t.Helper()
-	defer func() { _ = resp.Body.Close() }()
-	b, err := io.ReadAll(resp.Body)
-	if err != nil {
-		t.Fatalf("read body: %v", err)
-	}
-	return b
-}
